Keep StopFailure tests in hooks suite
